shell fixes etc
popups, scroll, logger etc
